ETCH-157 compiler: add @Signal annotation support
The @Signal method annotation causes a new Signal to be
created and added to the service. Two methods (subscribe and unsubscribe)
are generated and added to the signal. The subscription methods are
two-way so that they could be overriden and throw an exception if
the subscription fails for any reason.

Patch from Aleksandar Kanchev <kanchev@itestra.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/etch/branches/singlestack@1149209 13f79535-47bb-0310-9956-ffa450edef68