ETCH-157 binding-csharp: add @Signal support
All the signal management methods are generated with a "_" prefix
as "private". Only the methods with their original signature, which
were defined in the etch idl are exported within the client intf.
The rest are accessible through the base or remote classes.

The list of subscribed clients is managed within the client's remote
instance, in order to provide a default implmentation for sending a
signal to all subscribed clients.

Minimal changes were needed within the runtime in order to compare
WhoSender instances properly and to provide _Send() method with a
Who recipient.

Patch was from Aleksandar Kanchev <kanchev@itestra.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/etch/branches/singlestack@1149221 13f79535-47bb-0310-9956-ffa450edef68