ETCH-157 binding-csharp: add mono support
updated build-system to apache-ant-dotnet 1.1
updated build-files to avoid building/executing tests on mono (NUnit not working at the moment)

Adds support for building and using the C# binding with Mono (>= 2.10.1).
Changes the build.xml files to use xbuild instead of mdtool and bumps the dependency to Mono 2.10.1. Additional the HPTimer class is rewritten to use .NET methods and not Win32 methods.

This patch comes from
    Armin Mueller <mueller@itestra.com>
    Aleksandar Kanchev <kanchev@itestra.com>

git-svn-id: https://svn.apache.org/repos/asf/incubator/etch/branches/singlestack@1149260 13f79535-47bb-0310-9956-ffa450edef68