ETCH-280 Updating CAPU and adapting logger interface
This patch upgrades to CMake dependency to capu 0.13.0.
Due to this upgrade the logger interface needs to be changed
to use the new logging mechanism provided by CAPU 0.13.0.
In order to achieve this the EtchLogger has been extended by
a new macro called ETCH_LOG_<LOGLEVEL> in order to add an
abstraction from the CAPU logger.

Change-Id: I118008422c63d9406d006afe1193ba9d49c3213b

git-svn-id: https://svn.apache.org/repos/asf/etch/trunk@1548398 13f79535-47bb-0310-9956-ffa450edef68