ETCH-287 Fixing deadlock while calling methods on client side
On calls getResult(), hasResult(), getException() or hasException() on a
AsyncResult object the client waits until it gets notified from the
mailbox that a respective message has arrived.
In case the connection on socket level was broken before or during a
method call the AsyncResult object was not notified about this error
state and blocked forever.

Change-Id: I511e2fc60eb8b347a0d184cc02a7fd3ec937f2b6

git-svn-id: https://svn.apache.org/repos/asf/etch/trunk@1578869 13f79535-47bb-0310-9956-ffa450edef68